changed floats to doubles

+++ FsFastDtw/Dtw.fs
@@ -9,7 +9,7 @@ type Dtw () =
   /// series1: First series
   /// series2: Second series
   /// Returns: distance between series
-  static member Distance (series1: float[][]) (series2: float[][]) :float =
+  static member Distance (series1: double[][]) (series2: double[][]) :double =
     let (cost, _) = dtw series1 series2 None
     cost
 
@@ -18,5 +18,5 @@ type Dtw () =
   /// series2: Second series
   /// radius: Search radius
   /// Returns: (distance between series, correlated point maps)
-  static member DistanceWithPath (series1: float[][]) (series2: float[][]) (radius: int) :(float * Point list) =
+  static member DistanceWithPath (series1: double[][]) (series2: double[][]) (radius: int) :(double * Point list) =
     dtw series1 series2 None
